refactor: remove regularDeleted references and optimize extraordinary payment filtering logic
