+++ src/types/hooks/IUseExtraordinaryCyclesForm/index.ts
@@ -10,7 +10,6 @@ interface IUseExtraordinaryCyclesForm {
   onSubmit: ((values: IExtraordinaryCyclesEntry) => void) | undefined;
   onFormValid: React.Dispatch<React.SetStateAction<boolean>> | undefined;
   extraordinaryPayment: IExtraordinaryCyclesEntry[];
-  regularDeleted: IOrdinaryCyclesEntry[];
   setExtraordinaryPayment: React.Dispatch<
     React.SetStateAction<IExtraordinaryCyclesEntry[]>
   >;
